requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
choicesenum==0.7.0
Django==4.0.3
django-exiffield==3.0.0
django-taggit==2.1.0
jsonfield==3.1.0
Pillow==9.0.1
psycopg2==2.9.3
six==1.16.0
sqlparse==0.4.2
django-cleanup==6.0.0
google-cloud-vision==2.7.2
